requirements.txt
ctransformers
sentence-transformers
pinecone-client
flask
pypdf
python-dotenv
pinecone[grpc]
langchain-pinecone
langchain
langchain-community
langchain-huggingface
-e .
